requirements.txt
redis
opencv-python
numpy
pyyaml
ultralytics
shapely

paddleocr==2.5
paddlepaddle-gpu==3.1.1 -i https://www.paddlepaddle.org.cn/packages/stable/cu118/
paddleslim==2.2.2